Connected library, browse, and study to ViewCard.
One you click on a stack from library, browse, or study you will be taken to ViewCard. ViewCard now automatically generates the cards of the stack you selected. Each card has a term in the front and a definition in the back. There is also a display below that tells you what card you're on out of the entire stack.